# backend/doc/api.yaml
openapi: 3.0.0
info:
  title: Newspaper Service
  description: Daily news website service
  version: 1.0.0
servers:
  - url: http://localhost:4000/
components:
  schemas:
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Adabiyot va sanat'
        slug:
          type: string
          example: 'adabiyot-va-sanat'
          format: slug
...
paths:
  /admin/category:
    get:
      tags:
        - Admin Category
      summary: Get all categories
      description: Gets all categories
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags:
        - Admin Category
      summary: Create category
      description: Creates a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  description: The category's name
                slug:
                  type: string
                  format: slug
                  description: The category's slug
              required:
                - name
                - slug
      responses:
        201:
          description: The user has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Bad request. body does not contain required fields
        409:
          description: Conflict. Category slug already exists!
  /admin/category/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Admin Category
      summary: Get single category by id
      description: Get single category by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Bad request. categoryId is invalid
        404:
          description: Category not found
    put:
      tags:
        - Admin Category
      summary: Update category information
      description: Update category information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  example: Adabiyot va sanat
                slug:
                  type: string
                  format: slug
                  example: adabiyot-va-sanat
              required:
                - slug
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Bad request. categoryId is invalid
        404:
          description: Category not found
    delete:
      tags:
        - Admin Category
      summary: Delete category
      description: Delete category from service
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. categoryId is invalid (not uuid)
        404:
          description: Category was not found.
...